feat: moved router to src folder
