--- src/client/router/index.ts
@@ -1,22 +1,0 @@
-import {createRouter, createWebHistory, RouteRecordRaw} from "vue-router";
-import Auth from "@/client/src/components/Auth/Auth.vue"
-
-const routes: Array<RouteRecordRaw> = [
-    {
-        path: "/",
-        name: "Auth",
-        component: Auth
-    },
-    {
-        path: "/auth",
-        name: "Auth",
-        component: Auth
-    }
-]
-
-const router = createRouter({
-    history: createWebHistory(process.env.BASE_URL),
-    routes
-})
-
-export default router
